+++ truffle/src/com.oracle.truffle.api.vm/src/com/oracle/truffle/api/vm/PolyglotProxy.java
@@ -438,24 +438,77 @@ final class PolyglotProxy {
 
     }
 
+    private static final class ProxyRemoveNode extends ProxyRootNode {
+
+        @Override
+        Object executeProxy(PolyglotLanguageContext context, Proxy proxy, Object[] arguments) {
+            if (arguments.length >= 2) {
+                Object key = arguments[1];
+                if (proxy instanceof ProxyArray && key instanceof Number) {
+                    return removeArrayElement((ProxyArray) proxy, (Number) key);
+                } else if (proxy instanceof ProxyObject && key instanceof String) {
+                    return removeMember((ProxyObject) proxy, (String) key);
+                }
+            }
+            CompilerDirectives.transferToInterpreter();
+            throw UnsupportedMessageException.raise(Message.READ);
+        }
+
+        @TruffleBoundary
+        static boolean removeMember(ProxyObject object, String key) {
+            if (object.hasMember(key)) {
+                try {
+                    return object.removeMember(key);
+                } catch (UnsupportedOperationException e) {
+                    throw UnsupportedMessageException.raise(Message.READ);
+                }
+            } else {
+                throw UnknownIdentifierException.raise(key);
+            }
+        }
+
+        @TruffleBoundary
+        static boolean removeArrayElement(ProxyArray object, Number index) {
+            boolean result;
+            try {
+                result = object.remove(index.longValue());
+            } catch (ArrayIndexOutOfBoundsException e) {
+                throw UnknownIdentifierException.raise(e.getMessage());
+            } catch (UnsupportedOperationException e) {
+                throw UnsupportedMessageException.raise(Message.READ);
+            }
+            return result;
+        }
+
+    }
+
     private static final class EngineProxyFactory implements StandardFactory {
 
         private static final ForeignAccess INSTANCE = ForeignAccess.create(EngineProxy.class, new EngineProxyFactory());
 
+        @Override
         public CallTarget accessWrite() {
             return Truffle.getRuntime().createCallTarget(new ProxyWriteNode());
         }
 
+        @Override
         public CallTarget accessIsBoxed() {
             return Truffle.getRuntime().createCallTarget(new ProxyIsBoxedNode());
         }
 
+        @Override
         public CallTarget accessUnbox() {
             return Truffle.getRuntime().createCallTarget(new ProxyUnboxNode());
         }
 
+        @Override
         public CallTarget accessRead() {
             return Truffle.getRuntime().createCallTarget(new ProxyReadNode());
+        }
+
+        @Override
+        public CallTarget accessRemove() {
+            return Truffle.getRuntime().createCallTarget(new ProxyRemoveNode());
         }
 
         @Override
@@ -463,6 +516,7 @@ final class PolyglotProxy {
             return Truffle.getRuntime().createCallTarget(new ProxyIsInstantiableNode());
         }
 
+        @Override
         public CallTarget accessNew(int argumentsLength) {
             return Truffle.getRuntime().createCallTarget(new ProxyNewNode());
         }
@@ -472,42 +526,52 @@ final class PolyglotProxy {
             return Truffle.getRuntime().createCallTarget(new ProxyHasKeysNode());
         }
 
+        @Override
         public CallTarget accessKeys() {
             return Truffle.getRuntime().createCallTarget(new ProxyKeysNode());
         }
 
+        @Override
         public CallTarget accessKeyInfo() {
             return Truffle.getRuntime().createCallTarget(new ProxyKeyInfoNode());
         }
 
+        @Override
         public CallTarget accessIsNull() {
             return Truffle.getRuntime().createCallTarget(RootNode.createConstantNode(false));
         }
 
+        @Override
         public CallTarget accessIsExecutable() {
             return Truffle.getRuntime().createCallTarget(new ProxyIsExecutableNode());
         }
 
+        @Override
         public CallTarget accessInvoke(int argumentsLength) {
             return Truffle.getRuntime().createCallTarget(new ProxyInvokeNode());
         }
 
+        @Override
         public CallTarget accessHasSize() {
             return Truffle.getRuntime().createCallTarget(new ProxyHasSizeNode());
         }
 
+        @Override
         public CallTarget accessGetSize() {
             return Truffle.getRuntime().createCallTarget(new ProxyGetSizeNode());
         }
 
+        @Override
         public CallTarget accessExecute(int argumentsLength) {
             return Truffle.getRuntime().createCallTarget(new ProxyExecuteNode());
         }
 
+        @Override
         public CallTarget accessIsPointer() {
             return Truffle.getRuntime().createCallTarget(new ProxyIsPointerNode());
         }
 
+        @Override
         public CallTarget accessAsPointer() {
             return Truffle.getRuntime().createCallTarget(new ProxyAsPointerNode());
         }
@@ -517,6 +581,7 @@ final class PolyglotProxy {
             return null;
         }
 
+        @Override
         public CallTarget accessMessage(Message unknown) {
             return null;
         }